backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.7.2
networkx==3.3
easyocr==1.7.1
PyPDF2==3.0.1
python-multipart==0.0.9
openai==1.65.1
torch==2.1.0
torch-geometric==2.7.0
faiss-cpu==1.13.2
spacy==3.8.13
reportlab==4.4.10
